requirements.txt
fastapi~=0.115.12
uvicorn
python-socketio~=5.13.0

aiogram~=3.20.0.post0